vertex_trigger_function_job2/requirements.txt
# requirements.txt (for Job 2 Cloud Function)
functions-framework==3.*
google-cloud-aiplatform>=1.25.0
google-cloud-storage>=2.0.0
google-cloud-secret-manager>=2.16.0